ASU-1712: Dont delete after-applications as they are replaced instead

+++ public/modules/custom/asu_application/src/Form/ApplicationForm.php
@@ -533,6 +533,8 @@ HTML;
 
   /**
    * Handle saving the form values.
+   * 
+   * Deletes old application.
    *
    * @param array $form
    *   Form array.
@@ -548,7 +550,12 @@ HTML;
     $oldBackendId = $this->entity->get('field_backend_id')->value ?? NULL;
     $confirmDeletion = $form_state->getValue('confirm_application_deletion') ?? 'NOT SET';
 
-    if ($oldBackendId && $confirmDeletion == '1') {
+    $project_id = $this->entity->get('project_id')->value;
+    $project = $this->entityTypeManager->getStorage('node')->load($project_id);
+    $canApplyAfterwards = $project->get('field_can_apply_afterwards')->value;
+
+    // applications made after shouldn't be deleted
+    if ($oldBackendId && $confirmDeletion == '1' && !$canApplyAfterwards) {
       try {
         $user = \Drupal::entityTypeManager()->getStorage('user')->load(\Drupal::currentUser()->id());
         \Drupal::service('asu_api.backendapi')->deleteApplication($user, $oldBackendId);
